fix validator
