+++ jnosql-mapping/jnosql-mapping-validation/src/test/java/org/eclipse/jnosql/mapping/validation/ColumnMockProducer.java
@@ -14,16 +14,16 @@
  */
 package org.eclipse.jnosql.mapping.validation;
 
-import jakarta.nosql.column.Column;
-import jakarta.nosql.column.ColumnEntity;
-import jakarta.nosql.column.ColumnManager;
-import org.mockito.Mockito;
-
 import jakarta.annotation.Priority;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.enterprise.inject.Alternative;
 import jakarta.enterprise.inject.Produces;
 import jakarta.interceptor.Interceptor;
+import org.eclipse.jnosql.communication.column.Column;
+import org.eclipse.jnosql.communication.column.ColumnEntity;
+import org.eclipse.jnosql.communication.column.ColumnManager;
+import org.mockito.Mockito;
+
 import java.math.BigDecimal;
 import java.util.function.Supplier;
