+++ jnosql-mapping/jnosql-mapping-validation/src/test/java/org/eclipse/jnosql/mapping/validation/DocumentTemplateValidationTest.java
@@ -14,7 +14,7 @@
  */
 package org.eclipse.jnosql.mapping.validation;
 
-import jakarta.nosql.mapping.document.DocumentTemplate;
+import jakarta.nosql.document.DocumentTemplate;
 import jakarta.nosql.tck.test.CDIExtension;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
