+++ jnosql-mapping/jnosql-mapping-validation/src/test/java/org/eclipse/jnosql/mapping/validation/Person.java
@@ -15,9 +15,9 @@
 package org.eclipse.jnosql.mapping.validation;
 
 
-import jakarta.nosql.mapping.Column;
-import jakarta.nosql.mapping.Entity;
-import jakarta.nosql.mapping.Id;
+import jakarta.nosql.Column;
+import jakarta.nosql.Entity;
+import jakarta.nosql.Id;
 
 import jakarta.validation.constraints.DecimalMax;
 import jakarta.validation.constraints.Min;
